+++ NEWS.md
@@ -8,5 +8,6 @@
     * Tidied up and simplified code.
 * Updated `run.umap` to overwrite `umap.x.name` and `umap.y.name` columns in 
 input data.table if they already exist.
+* Updated `do.asinh` to overwrite columns in input data.table if they already exist.
 * Added a `NEWS.md` file to track changes to the package.
 * Added `testthat` unittests.